Add images

+++ DX11Raytracer/DX11Raytracer/src/IceMaterial.cpp
@@ -22,7 +22,7 @@ namespace gfx
 		const double n0 = PerlinNoise::GetNoise3D(rec.positionWS * vec3(55.0, 1.0, 55.0), 7u);
 		const double roughness = std::pow(Saturate(n0), 3.0) * 0.15;
 
-		attenuation = Color(1.0f) * PerlinNoise::GetNoise3D(rec.positionWS, 2u);
+		attenuation = Color(1.0f) * static_cast<float>(PerlinNoise::GetNoise3D(rec.positionWS, 2u));
 
 		// This assumes the other medium is air - need to update if adding more complicated situations
 		const double refractionRatio = rec.isFrontFacing ? (1.0 / 1.33) : 1.33; // should be 1.33, but this looks better...
